Fix catastrophic bug in sfsck, found by Jim Danz.
The number of inodes it remembered was never incremented, so none of
the inode checking ever did a damn thing.

Fixing that exposed another glaring problem: there's an ifdef around
the array reallocation, for with and without realloc, and one of the
branches allocates a block for the wrong size of object. So, fix that
too.

Sigh. This code has been in use for how long?